Moved MetaStoreDirectSql, StatObjectConverter, and ColumnStatsAggregator and implementations.  Combined all of these together because there was a circular dependence.
