+++ standalone-metastore/src/main/java/org/apache/hadoop/hive/metastore/columnstats/aggr/BinaryColumnStatsAggregator.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -21,19 +21,19 @@ package org.apache.hadoop.hive.metastore.columnstats.aggr;
 
 import java.util.List;
 
-import org.apache.hadoop.hive.metastore.api.BooleanColumnStatsData;
+import org.apache.hadoop.hive.metastore.api.BinaryColumnStatsData;
 import org.apache.hadoop.hive.metastore.api.ColumnStatistics;
 import org.apache.hadoop.hive.metastore.api.ColumnStatisticsData;
 import org.apache.hadoop.hive.metastore.api.ColumnStatisticsObj;
 import org.apache.hadoop.hive.metastore.api.MetaException;
 
-public class BooleanColumnStatsAggregator extends ColumnStatsAggregator {
+public class BinaryColumnStatsAggregator extends ColumnStatsAggregator {
 
   @Override
   public ColumnStatisticsObj aggregate(String colName, List<String> partNames,
       List<ColumnStatistics> css) throws MetaException {
     ColumnStatisticsObj statsObj = null;
-    BooleanColumnStatsData aggregateData = null;
+    BinaryColumnStatsData aggregateData = null;
     String colType = null;
     for (ColumnStatistics cs : css) {
       if (cs.getStatsObjSize() != 1) {
@@ -47,19 +47,18 @@ public class BooleanColumnStatsAggregator extends ColumnStatsAggregator {
         statsObj = ColumnStatsAggregatorFactory.newColumnStaticsObj(colName, colType, cso
             .getStatsData().getSetField());
       }
-      BooleanColumnStatsData newData = cso.getStatsData().getBooleanStats();
+      BinaryColumnStatsData newData = cso.getStatsData().getBinaryStats();
       if (aggregateData == null) {
         aggregateData = newData.deepCopy();
       } else {
-        aggregateData.setNumTrues(aggregateData.getNumTrues() + newData.getNumTrues());
-        aggregateData.setNumFalses(aggregateData.getNumFalses() + newData.getNumFalses());
+        aggregateData.setMaxColLen(Math.max(aggregateData.getMaxColLen(), newData.getMaxColLen()));
+        aggregateData.setAvgColLen(Math.max(aggregateData.getAvgColLen(), newData.getAvgColLen()));
         aggregateData.setNumNulls(aggregateData.getNumNulls() + newData.getNumNulls());
       }
     }
     ColumnStatisticsData columnStatisticsData = new ColumnStatisticsData();
-    columnStatisticsData.setBooleanStats(aggregateData);
+    columnStatisticsData.setBinaryStats(aggregateData);
     statsObj.setStatsData(columnStatisticsData);
     return statsObj;
   }
-
 }
