+++ standalone-metastore/src/test/java/org/apache/hadoop/hive/metastore/TestDeadline.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
